Decoupling from postgresql-jdbc https://github.com/postgis/postgis-java/issues/18
- Established geometry module and began moving geometry classes to new module

- Moved geometry parsing/building code out of PGgeometry and into separate GeometryBuilder class
  in the postgis-geometry module and refactored all code to use new GeometryBuilder class instead
  of PGgeometry

- Moved binary and util packages from postgis-jdbc module to postgis-geometry
