+++ pom.xml
@@ -69,6 +69,7 @@
 
     <modules>
         <module>tools</module>
+        <module>postgis-geometry</module>
         <module>jdbc</module>
         <module>jdbc_jtsparser</module>
         <module>postgis-jdbc-java2d</module>
