+++ postgis-geometry/src/main/java/org/postgis/ComposedGeom.java
