+++ postgis-geometry/src/main/java/org/postgis/GeometryTokenizer.java
