+++ postgis-geometry/src/main/java/org/postgis/LineString.java
