+++ postgis-geometry/src/main/java/org/postgis/LinearRing.java
