+++ postgis-geometry/src/main/java/org/postgis/MultiLineString.java
