+++ postgis-geometry/src/main/java/org/postgis/MultiPoint.java
