+++ postgis-geometry/src/main/java/org/postgis/Point.java
