+++ postgis-geometry/src/main/java/org/postgis/PointComposedGeom.java
