+++ postgis-geometry/src/main/java/org/postgis/binary/ByteGetter.java
