+++ postgis-geometry/src/main/java/org/postgis/binary/ByteSetter.java
