+++ postgis-geometry/src/main/java/org/postgis/binary/ValueGetter.java
