+++ postgis-geometry/src/main/java/org/postgis/binary/ValueSetter.java
