+++ postgis-geometry/src/main/java/org/postgis/util/VersionFunctions.java
