+++ postgis-geometry/src/main/java/org/postgis/util/VersionUtil.java
