+++ postgis-geometry/src/main/java/org/postgis/util/package-info.java
