+++ postgis-geometry/src/test/java/org/postgis/TokenizerTest.java
