+++ postgis-geometry/src/test/resources/logback-test.xml
@@ -1,0 +1,19 @@
+<configuration debug="false">
+
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <layout class="ch.qos.logback.classic.PatternLayout">
+            <Pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</Pattern>
+        </layout>
+    </appender>
+
+    <root level="debug">
+        <appender-ref ref="STDOUT" />
+    </root>
+
+    <logger name="com.github.dockerjava" level="ERROR"/>
+    <logger name="org.testcontainers" level="ERROR"/>
+    <logger name="net.postgis" level="ERROR"/>
+
+    <logger name="org.postgis" level="ERROR"/>
+
+</configuration>
